fix: tests and renamed file
