+++ src/modules/checkout/verify.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 
-import { ChainZod, PackageIdZod, UrlZod } from "../../utils/validators";
+import { ChainZod, PackageIdZod, UrlZod } from "../../utils/zod";
 
 import { InvalidCheckoutSessionInputError } from "./errors";
 import { ICheckoutSessionInput } from "./types";
